+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 - fromBase10Conversion
+- toBase10HornerConversion
+- guidelines to use tester.m
 
 ## 0.0 - 2017-03-08